+++ src/GRA.Domain.Repository/IUserRepository.cs
@@ -21,5 +21,6 @@ namespace GRA.Domain.Repository
             PageHouseholdAsync(int householdHeadUserId, int skip, int take);
         Task SetUserPasswordAsync(int currentUserId, int userId, string password);
         Task<IEnumerable<User>> GetHouseholdAsync(int householdHeadUserId);
+        Task<bool> UsernameInUseAsync(int siteId, string username);
     }
 }